refactor : order model-domain 분리
