+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/order/dto/OrderAddRequestDto.java
@@ -1,6 +1,6 @@
 package com.sparta.blackwhitedeliverydriver.order.dto;
 
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderTypeEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderTypeEnum;
 import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
