+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/order/dto/OrderUpdateRequestDto.java
@@ -1,6 +1,6 @@
 package com.sparta.blackwhitedeliverydriver.order.dto;
 
-import com.sparta.blackwhitedeliverydriver.order.entity.OrderStatusEnum;
+import com.sparta.blackwhitedeliverydriver.model.order.entity.OrderStatusEnum;
 import jakarta.validation.constraints.NotNull;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
